Implement advanced debug outputs and functionality improvements
Improved the handling and output of debug information by introducing message queues in the ConversableAgent. This enables detailed tracking and output of responses during testing. Modified the logic around the generation and receipt of messages within the ConversableAgent class, ensuring that repeated or identical messages are handled more effectively. In the unlimited_gpt and 5_build_agents modules, added new operations and settings to enhance functionality and user experience.

1. Supported building agents, skills and workflows in web UI.
2. Supported 2 roles playground.
3. Enhanced the ChatManager to support multi-tasks.
4. Fixed the error of unexecutable code extraction.
5. Simplified the config list fetching from json file only.
6. Supported multithread message communication on the conversible agent level.
7. Fixed the issue of duplicate historical messages.
8. Fixed the incorrect last replied message index and the issue of missing system message.
9. Fixed the markdown format issue for table.
10. Fixed the system message prompt issue.